+++ lib/src/extension.dart
@@ -1,12 +1,14 @@
 import 'dart:math' as math;
 
-import 'light_stats.dart';
-import 'stats.dart';
+import '../stats.dart';
 
 extension StatsExtensions<T extends num> on Iterable<T> {
   Stats<T> get stats => Stats<T>.fromData(this);
 
   LightStats<T> get lightStats => LightStats<T>.fromData(this);
+
+  ConfidenceInterval confidenceInterval(ConfidenceLevel confidenceLevel) =>
+      ConfidenceInterval.calculate(cast<num>(), confidenceLevel);
 
   /// Returns the maximum of all values in `this`.
   T get max => reduce(math.max);