+++ lib/stats.dart
@@ -1,3 +1,5 @@
+export 'src/confidence_interval.dart' show ConfidenceInterval;
+export 'src/confidence_level.dart' show ConfidenceLevel;
 export 'src/extension.dart';
 export 'src/light_stats.dart';
 export 'src/stats.dart';